Fixed leave group socket bug
A dependency in useEffect caused the page to reload every time a
socket event was emitted. Removing that dependency caused this bug
to appear. To fix it, we need to check if deleting a group caused
the index not to change. If the index remains the same, we re-fetch
the photo list.